refactor: rename field values file
